Add files via upload
working prototype. not all functies do stuff but it has a working menu fixed to the bottom. 
file contains the WWW folder of phonegap. i wil also upload the zip file if u want to try it out for yourself. u can upload that zipfile on the website of phonegab build.